new version with resave

+++ src/net/preibisch/ijannot/util/IOFunctions.java
@@ -1,10 +1,14 @@
 package net.preibisch.ijannot.util;
 
+import java.io.BufferedReader;
 import java.io.File;
+import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.file.Files;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import javax.swing.SwingUtilities;
 
@@ -25,18 +29,19 @@ public class IOFunctions {
 			System.out.println(string);
 	}
 
-	public static <T> void generateCSV(List<T> list, File file) {
+	public static <T> void generateCSV(List<T> list, File file, Boolean resave) {
 		println("creating CSV: " + file.getAbsolutePath());
-		if (file.exists()) {
-			try {
-				Files.delete(file.toPath());
-			} catch (IOException e) {
-				Log.error(e.toString());
-				return;
+		if (resave)
+			if (file.exists()) {
+				try {
+					Files.delete(file.toPath());
+				} catch (IOException e) {
+					Log.error(e.toString());
+					return;
+				}
 			}
-		}
 
-		try (FileWriter csvWriter = new FileWriter(file)) {
+		try (FileWriter csvWriter = new FileWriter(file, true)) {
 
 			for (Object a : list) {
 				String s = a.toString();
@@ -58,6 +63,25 @@ public class IOFunctions {
 		} catch (IOException e) {
 			Log.error("Can't delete folder : " + e.toString());
 		}
+	}
+
+	public static Map<String, Integer> getCSV(File f) throws IOException {
+		Map<String, Integer> result = new HashMap<>();
+
+		BufferedReader csvReader = new BufferedReader(new FileReader(f));
+		String row;
+		while ((row = csvReader.readLine()) != null) {
+			String[] data = row.split(",");
+			String key = data[0].substring(1, data[0].length() - 1);
+			// System.out.println(key);
+			result.put(key, Integer.parseInt(data[1]));
+		}
+		csvReader.close();
+		return result;
+	}
+
+	public static void main(String[] args) throws IOException {
+		getCSV(new File("/Users/Marwan/Desktop/block_annotation.csv"));
 	}
 
 }
